refac(onBoarding): refactoring and edit handle container press
