+++ src/features/onBoarding/constant.js
@@ -1,0 +1,41 @@
+import OnBoarding1 from "../../../assets/onBoarding1.svg";
+import OnBoarding2 from "../../../assets/onBoarding2.svg";
+import OnBoarding3 from "../../../assets/onBoarding3.svg";
+import OnBoarding4 from "../../../assets/onBoarding4.svg";
+import OnBoardingProgress1 from "../../../assets/onBoardingProgress1.svg";
+import OnBoardingProgress2 from "../../../assets/onBoardingProgress2.svg";
+import OnBoardingProgress3 from "../../../assets/onBoardingProgress3.svg";
+import OnBoardingProgress4 from "../../../assets/onBoardingProgress4.svg";
+
+export const lastIndex = 4;
+
+export const onBoardingList = [
+  <OnBoarding1 />,
+  <OnBoarding2 />,
+  <OnBoarding3 height={240} />,
+  <OnBoarding4 />,
+];
+export const onBoardingTitleText = [
+  "폴더별로 관리",
+  "쉽고 빠르게 저장",
+  "편리한 검색",
+  "복사 및 공유",
+];
+export const onBoardingTextTop = [
+  "링크를 유형별로 분류하여 ",
+  "복사한 URL만 붙여 넣으면 ",
+  "저장한 이름과 URL을 검색해서 ",
+  "저장한 링크를 복사하거나, ",
+];
+export const onBoardingTextBottom = [
+  "깔끔하게 관리할 수 있어요!",
+  "링크 이름과 썸네일을 자동으로 추출해 줘요.",
+  "원하는 링크를 빠르게 찾을 수 있어요.",
+  "다른 앱으로 공유할 수 있어요.",
+];
+export const onBoardingProgressList = [
+  <OnBoardingProgress1 />,
+  <OnBoardingProgress2 />,
+  <OnBoardingProgress3 />,
+  <OnBoardingProgress4 />,
+];
